Fixing Issue with Role rm Command (#96)
Fixing an issue with the command 'gzr role rm ROLE_ID' where ROLE_ID is the given ID of the role to remove/delete. The internal 'detele_role()' command was using an undefined variable instead of using the role_id passed in as the ID parameter. However, the internal 'delete_role()' command was not throwing any errors/exceptions when called with a null ID parameter, so it appears that the command was successful when it was not.

Co-authored-by: Todd Kofford <tdkoff@gmail.com>